+++ src/store/index.js
@@ -1,5 +1,10 @@
-import { createStore } from "redux";
-import { profileReducer } from "./profile/reducer";
+import { combineReducers, createStore } from "redux";
+import chatsReducer from "./chats/chatsReducer";
+import profileReducer from "./profile/reducerProfile";
 
-export const store = createStore(profileReducer, window.__REDUX_DEVTOOLS_EXTENSION__ &&
+const rootReducer = combineReducers({
+    profile: profileReducer,
+    chats: chatsReducer,
+});
+export const store = createStore(rootReducer, window.__REDUX_DEVTOOLS_EXTENSION__ &&
     window.__REDUX_DEVTOOLS_EXTENSION__());
