+++ src/store/types.js
@@ -6,3 +6,6 @@ export const SET_USER_NAME = 'SET_USER_NAME';
 export const ADD_CHAT = 'ADD_CHAT';
 export const REMOVE_CHAT = 'REMOVE_CHAT';
 
+//Types for messages
+export const ADD_NEW_MESSAGE = 'ADD_MESSAGE';
+
